Move files around; add reusable Page component
