+++ client/src/app/App.css
@@ -1,6 +1,6 @@
-@import "../node_modules/normalize.css/normalize.css";
-@import "../node_modules/@blueprintjs/core/lib/css/blueprint.css";
-@import "../node_modules/@blueprintjs/icons/lib/css/blueprint-icons.css";
+@import "../../node_modules/normalize.css/normalize.css";
+@import "../../node_modules/@blueprintjs/core/lib/css/blueprint.css";
+@import "../../node_modules/@blueprintjs/icons/lib/css/blueprint-icons.css";
 
 html, body, #root {
   height: 100%;
