+++ client/src/stories/StoriesPage.tsx
@@ -1,0 +1,11 @@
+import StoriesFilter from './StoriesFilter';
+import React from 'react';
+import StoriesList from './StoriesList';
+import Page from '../app/Page';
+
+export default () => (
+  <Page>
+    <StoriesFilter />
+    <StoriesList />
+  </Page>
+);
